+++ app/scripts/com/2fdevs/videogular/plugins/buffering.js
@@ -6,24 +6,14 @@
 "use strict";
 /**
  * @ngdoc directive
- * @name com.2fdevs.videogular.plugins.buffering:buffering
+ * @name com.2fdevs.videogular.plugins.buffering:vgBuffering
  * @restrict E
  * @description
  * Shows a spinner when Videogular is buffering or preparing the video player.
  *
  * ```html
- * <videogular vg-width="config.width"
- *        vg-height="config.height"
- *        vg-theme="config.theme.url"
- *        vg-autoplay="config.autoPlay"
- *        vg-stretch="config.stretch.value"
- *        vg-responsive="config.responsive">
- *    <video preload='metadata'>
- *        <source src="assets/videos/videogular.mp4" type="video/mp4">
- *        <source src="assets/videos/videogular.webm" type="video/webm">
- *
- *    <track kind="captions" src="assets/subs/pale-blue-dot.vtt" srclang="en" label="English" default></track>
- *    </video>
+ * <videogular vg-theme="config.theme.url" vg-autoplay="config.autoPlay">
+ *    <vg-video vg-src="sources"></vg-video>
  *
  *    <vg-buffering></vg-buffering>
  * </videogular>