+++ docs/js/docs-setup.js
@@ -6,21 +6,57 @@ NG_DOCS={
   "pages": [
     {
       "section": "api",
-      "id": "com.2fdevs.videogular.plugins.poster:poster",
-      "shortName": "poster",
+      "id": "com.2fdevs.videogular.plugins.buffering:vgBuffering",
+      "shortName": "vgBuffering",
       "type": "directive",
       "moduleName": "com.2fdevs.videogular.plugins",
-      "shortDescription": "Shows an image when player hasn&#39;t been played or has been completed a video.",
-      "keywords": "api assets attributes autoplay captions completed config customize default directive en english hasn height html image kind label metadata mp4 played player plugins poster preload responsive src srclang stretch theme type url vg-autoplay vg-height vg-responsive vg-stretch vg-theme vg-url vg-width vgposterimage video videogular vtt webm width"
+      "shortDescription": "Shows a spinner when Videogular is buffering or preparing the video player.",
+      "keywords": "api autoplay buffering config directive html player plugins preparing sources spinner theme url vg-autoplay vg-src vg-theme video videogular"
     },
     {
       "section": "api",
-      "id": "com.2fdevs.videogular.videogular:videogular",
-      "shortName": "videogular",
+      "id": "com.2fdevs.videogular.plugins.overlayplay:vgOverlayPlay",
+      "shortName": "vgOverlayPlay",
+      "type": "directive",
+      "moduleName": "com.2fdevs.videogular.plugins",
+      "shortDescription": "Shows a big play button centered when player is paused or stopped.",
+      "keywords": "api autoplay big button centered config directive html overlayplay paused play player plugins sources stopped theme url vg-autoplay vg-src vg-theme videogular"
+    },
+    {
+      "section": "api",
+      "id": "com.2fdevs.videogular.plugins.poster:vgPoster",
+      "shortName": "vgPoster",
+      "type": "directive",
+      "moduleName": "com.2fdevs.videogular.plugins",
+      "shortDescription": "Shows an image when player hasn&#39;t been played or has been completed a video.",
+      "keywords": "api attributes autoplay completed config customize directive hasn html image played player plugins poster sources theme url vg-autoplay vg-src vg-theme vg-url vgposterimage video videogular"
+    },
+    {
+      "section": "api",
+      "id": "com.2fdevs.videogular.vgAudio",
+      "shortName": "com.2fdevs.videogular.vgAudio",
       "type": "directive",
       "moduleName": "com.2fdevs.videogular",
-      "shortDescription": "Main directive that must wrap a &lt;video&gt; tag and all plugins.",
-      "keywords": "allow api appear area aspect attributes auto bars binding black boolean calculated call callback case centered change changes click completed consuming container content controller controls cover create css current customize data declared devices directive disabled duration enter entire exists fill fit free function header height image initialized inject inside layer leaving link main metadata milliseconds mobile mode modes mp4 number original param parameter params passing pause plans play playback player playing plugin plugins preserve prevent quality ratio receives representing required responsive scale scope sd set sets size source sources start stretch string tag tags time true updated updating user values variable vgautoplay vgchangesource vgcomplete vgheight vgplayerready vgresponsive vgstretch vgtheme vgupdatesize vgupdatestate vgupdatetime vgupdatevolume vgwidth video videogular volume width wrap"
+      "shortDescription": "Directive to add a source of audios. This directive will create a &lt;audio&gt; tag and usually will be above plugin tags.",
+      "keywords": "$sce add api array assets attributes audio audios auto bindable boolean click content controls create customize data default devices directive disabled en english http initialized kind label list mobile mp4 native object parameter playing plugin preload properties property required resource scope set source sources src srclang start string subtitles tag tags track true trusful trustasresourceurl type url user variable vgaudio vgloop vgnativecontrols vgpreload vgsrc vgtracks video videogular visible vtt"
+    },
+    {
+      "section": "api",
+      "id": "com.2fdevs.videogular.vgVideo",
+      "shortName": "com.2fdevs.videogular.vgVideo",
+      "type": "directive",
+      "moduleName": "com.2fdevs.videogular",
+      "shortDescription": "Directive to add a source of videos. This directive will create a &lt;video&gt; tag and usually will be above plugin tags.",
+      "keywords": "$sce add api array assets attributes auto bindable boolean click content controls create customize data default devices directive disabled en english http initialized kind label list mobile mp4 native object parameter playing plugin preload properties property required resource scope set source sources src srclang start string subtitles tag tags track true trusful trustasresourceurl type url user variable vgloop vgnativecontrols vgpreload vgsrc vgtracks vgvideo video videogular videos visible vtt"
+    },
+    {
+      "section": "api",
+      "id": "com.2fdevs.videogular.videogular",
+      "shortName": "com.2fdevs.videogular.videogular",
+      "type": "directive",
+      "moduleName": "com.2fdevs.videogular",
+      "shortDescription": "Main directive that must wrap a &lt;vg-video&gt; or &lt;vg-audio&gt; tag and all plugins.",
+      "keywords": "allow api attributes auto boolean call callback change changes click completed consuming content controller css current customize data devices directive disabled duration free function header initialized inject layer link main milliseconds mobile mp4 param parameter params pause plans play playback playing plugin plugins prevent quality receives required scope sd source sources start string tag tags time updated updating user values variable vgautoplay vgchangesource vgcomplete vgplayerready vgtheme vgupdatestate vgupdatetime vgupdatevolume video videogular volume wrap"
     }
   ],
   "apis": {
@@ -28,6 +64,7 @@ NG_DOCS={
     "guide": false
   },
   "html5Mode": true,
+  "editExample": true,
   "startPage": "/api",
   "scripts": [
     "angular.min.js"